appreparto-back/src/Front/requirements.txt
altair==4.2.2
asttokens==2.2.1
attrs==22.2.0
backcall==0.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.2
blinker==1.5
branca==0.6.0
bs4==0.0.1
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
comm==0.1.2
contourpy==1.0.7
cramjam==2.6.2
cycler==0.11.0
debugpy==1.6.6
decorator==5.1.1
entrypoints==0.4
executing==1.2.0
fastparquet==2023.2.0
folium==0.14.0
fonttools==4.39.0
fsspec==2023.3.0
future==0.18.3
geocoder==1.38.1
geographiclib==2.0
geopy==2.3.0
gitdb==4.0.10
GitPython==3.1.31
idna==3.4
importlib-metadata==6.0.0
importlib-resources==5.12.0
ipykernel==6.21.3
ipython==8.11.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
jupyter_client==8.0.3
jupyter_core==5.2.0
kiwisolver==1.4.4
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdurl==0.1.2
nest-asyncio==1.5.6
numpy==1.24.2
packaging==23.0
pandas==1.5.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
pkgutil_resolve_name==1.3.10
platformdirs==3.1.1
prompt-toolkit==3.0.38
protobuf==3.20.3
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==11.0.0
pydeck==0.8.0
Pygments==2.14.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
pyzmq==25.0.0
ratelim==0.1.6
requests==2.28.2
rich==13.3.2
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
semver==2.13.0
six==1.16.0
smmap==5.0.0
soupsieve==2.4
stack-data==0.6.2
streamlit==1.20.0
streamlit-folium==0.11.1
streamlit-option-menu==0.3.2
threadpoolctl==3.1.0
toml==0.10.2
toolz==0.12.0
tornado==6.2
traitlets==5.9.0
typing_extensions==4.5.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.15
validators==0.20.0
watchdog==2.3.1
wcwidth==0.2.6
zipp==3.15.0
psycopg2-binary == 2.9.5
python-dotenv == 1.0.0
